Fix handling of -- termination
Fix handling of -- termination. Previously, if you ran

$command -flag1 -etc -- some args after

the command's args would be doubled (some args after some args after).
Additionally, if there were items after the -- termination that looked
like flags, the parser would throw an error. It's better to terminate
parsing when -- is seen, because that usually indicates the caller wants
to pass the remainder of the arguments (including flags) as-is to
something else.