Add functions to support golang flags in pflags
They aren't "perfect".  spf13/pflag chose to use an explicit `Type` in the
value. golang has nothing of the sort and has a rather crazy
IsBoolFlag() bit of nonsense for the one case that really makes sense to
have the explicit `Type`. Probably because adding interfaces really screws
people up. So supporting golang flags in pflags isn't super super
simple, we need to figure out how to handle our `Type` given only the
information available in a golang flag. This does it as best we can with
`reflect`.